# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/patients": {
            "get": {
                "description": "list patient entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List patient entities",
                "operationId": "list-patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Patient"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create patient",
                "operationId": "create-patient",
                "parameters": [
                    {
                        "description": "Patient entity",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Patient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "description": "get patient by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a patient entity by ID",
                "operationId": "get-patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Patient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update patient by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a patient entity by ID",
                "operationId": "update-patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient entity",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Patient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get patient by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a patient entity by ID",
                "operationId": "delete-patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.AppointmentResults": {
            "type": "object",
            "properties": {
                "addtimeAppoint": {
                    "description": "AddtimeAppoint holds the value of the \"addtimeAppoint\" field.",
                    "type": "string"
                },
                "addtimeSave": {
                    "description": "AddtimeSave holds the value of the \"addtimeSave\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AppointmentResultsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AppointmentResultsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.AppointmentResultsEdges": {
            "type": "object",
            "properties": {
                "appointmentResultsToDoctor": {
                    "description": "AppointmentResultsToDoctor holds the value of the appointmentResultsToDoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "appointmentResultsToNurse": {
                    "description": "AppointmentResultsToNurse holds the value of the appointmentResultsToNurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "appointmentResultsToPatient": {
                    "description": "AppointmentResultsToPatient holds the value of the appointmentResultsToPatient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "appointmentResultsToRoom": {
                    "description": "AppointmentResultsToRoom holds the value of the appointmentResultsToRoom edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Room"
                }
            }
        },
        "ent.BloodType": {
            "type": "object",
            "properties": {
                "bloodValue": {
                    "description": "BloodValue holds the value of the \"bloodValue\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BloodTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BloodTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Diagnosis": {
            "type": "object",
            "properties": {
                "Opinionresult": {
                    "description": "Opinionresult holds the value of the \"Opinionresult\" field.",
                    "type": "string"
                },
                "diagnosisDate": {
                    "description": "DiagnosisDate holds the value of the \"diagnosisDate\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DiagnosisQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DiagnosisEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "symptom": {
                    "description": "Symptom holds the value of the \"symptom\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DiagnosisEdges": {
            "type": "object",
            "properties": {
                "doctorName": {
                    "description": "DoctorName holds the value of the Doctor_name edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "type": {
                    "description": "Type holds the value of the type edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TreatmentType"
                }
            }
        },
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "doctorName": {
                    "description": "DoctorName holds the value of the \"doctorName\" field.",
                    "type": "string"
                },
                "doctorPassword": {
                    "description": "DoctorPassword holds the value of the \"doctorPassword\" field.",
                    "type": "string"
                },
                "doctorUsername": {
                    "description": "DoctorUsername holds the value of the \"doctorUsername\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Gender": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "genderValue": {
                    "description": "GenderValue holds the value of the \"genderValue\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.MedicalProcedure": {
            "type": "object",
            "properties": {
                "Addtime": {
                    "description": "Addtime holds the value of the \"Addtime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicalProcedureQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicalProcedureEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.MedicalProcedureEdges": {
            "type": "object",
            "properties": {
                "doctor": {
                    "description": "Doctor holds the value of the Doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "procedureType": {
                    "description": "ProcedureType holds the value of the ProcedureType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProcedureType"
                }
            }
        },
...
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurseName": {
                    "description": "NurseName holds the value of the \"nurseName\" field.",
                    "type": "string"
                },
                "nursePassword": {
                    "description": "NursePassword holds the value of the \"nursePassword\" field.",
                    "type": "string"
                },
                "nurseUsername": {
                    "description": "NurseUsername holds the value of the \"nurseUsername\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Patient": {
            "type": "object",
            "properties": {
                "drugAllergy": {
                    "description": "DrugAllergy holds the value of the \"drugAllergy\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "hospitalNumber": {
                    "description": "HospitalNumber holds the value of the \"hospitalNumber\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "patientName": {
                    "description": "PatientName holds the value of the \"patientName\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PatientDetail": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientDetailQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientDetailEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientDetailEdges": {
            "type": "object",
            "properties": {
                "bloodtype": {
                    "description": "Bloodtype holds the value of the bloodtype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BloodType"
                },
                "gender": {
                    "description": "Gender holds the value of the gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "prefix": {
                    "description": "Prefix holds the value of the prefix edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prefix"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "patientDetails": {
                    "description": "PatientDetails holds the value of the patient_details edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.PatientDetail"
                    }
                },
                "patientToAppointmentResults": {
                    "description": "PatientToAppointmentResults holds the value of the PatientToAppointmentResults edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.AppointmentResults"
                    }
                },
                "patientToDiagnosis": {
                    "description": "PatientToDiagnosis holds the value of the PatientToDiagnosis edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Diagnosis"
                    }
                },
                "patientToMedicalProcedure": {
                    "description": "PatientToMedicalProcedure holds the value of the PatientToMedicalProcedure edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.MedicalProcedure"
                    }
                },
                "patientToRightToTreatment": {
                    "description": "PatientToRightToTreatment holds the value of the PatientToRightToTreatment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.RightToTreatment"
                    }
                },
                "triageResult": {
                    "description": "TriageResult holds the value of the triageResult edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.TriageResult"
                    }
                }
            }
        },
        "ent.Prefix": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrefixQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrefixEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "prefixValue": {
                    "description": "PrefixValue holds the value of the \"prefixValue\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.RightToTreatment": {
            "type": "object",
            "properties": {
                "EndTime": {
                    "description": "EndTime holds the value of the \"EndTime\" field.",
                    "type": "string"
                },
                "StartTime": {
                    "description": "StartTime holds the value of the \"StartTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RightToTreatmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RightToTreatmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RightToTreatmentEdges": {
            "type": "object",
            "properties": {
                "hospital": {
                    "description": "Hospital holds the value of the Hospital edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Hospital"
                },
                "patient": {
                    "description": "Patient holds the value of the Patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "rightToTreatmentType": {
                    "description": "RightToTreatmentType holds the value of the RightToTreatmentType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RightToTreatmentType"
                }
            }
        },
...
        "ent.Room": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "roomName": {
                    "description": "RoomName holds the value of the \"roomName\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.TreatmentType": {
            "type": "object",
            "properties": {
                "Type": {
                    "description": "Type holds the value of the \"Type\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TreatmentTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TreatmentTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.TriageResult": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TriageResultQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TriageResultEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "symptom": {
                    "description": "Symptom holds the value of the \"symptom\" field.",
                    "type": "string"
                },
                "triageDate": {
                    "description": "TriageDate holds the value of the \"triageDate\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TriageResultEdges": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "nurse": {
                    "description": "Nurse holds the value of the nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "urgencyLevel": {
                    "description": "UrgencyLevel holds the value of the urgencyLevel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UrgencyLevel"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}